requirements.txt
gunicorn==20.0.4
Flask==1.1.1
Jinja2==2.10.3
Flask-MySQLdb==0.2.0
python-dotenv==0.10.3
